backend/requirements.txt
django>=5.1,<6
djangorestframework>=3.15,<4
django-cors-headers>=4.6,<5
django-filter>=24.3,<25
psycopg[binary]>=3.2,<4
gunicorn>=23.0,<24
whitenoise>=6.8,<7
dj-database-url>=2.3,<3
dnspython>=2.7,<3
